requirements.txt
torch
torchvision
torchaudio
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyterlab
notebook
opencv-python
PyYAMLpip
rootutils
insightface
tqdm
onnxruntime
adjustText